refactor: Add NONE option to RomajiCase enum and implement transformation logic
This commit adds a new option "NONE" to the RomajiCase enum in the constants.ts file. It also implements the transformation logic for the "NONE" case in the transformTextToAnotherMode function in the extension.ts file. This allows the selected text to be returned as is without any transformation when the "NONE" case is selected.

Closes #issue-number